#284 update widgets-core module
