+++ widgets/src/iosMain/kotlin/dev/icerock/moko/widgets/core/factory/BaseInputViewFactory.kt
@@ -10,7 +10,6 @@ import dev.icerock.moko.resources.desc.StringDesc
 import dev.icerock.moko.widgets.core.ViewBundle
 import dev.icerock.moko.widgets.core.ViewFactory
 import dev.icerock.moko.widgets.core.ViewFactoryContext
-import dev.icerock.moko.widgets.core.shouldChangeCharacters
 import dev.icerock.moko.widgets.core.style.view.MarginValues
 import dev.icerock.moko.widgets.core.style.view.WidgetSize
 import dev.icerock.moko.widgets.core.utils.bind
@@ -69,11 +68,12 @@ abstract class BaseInputViewFactory<V : UIView> : ViewFactory<InputWidget<out Wi
     ) {
         field.data.bind { newValue ->
             val currentText = textField.text.orEmpty()
-            val shouldApplyChange = textField.delegate?.shouldChangeCharacters(
-                textField = textField,
-                range = NSMakeRange(0.toULong(), currentText.length.toULong()),
-                text = newValue
-            ) ?: true
+            val shouldApplyChange = textField.delegate?.run {
+                textField(
+                    textField,
+                    NSMakeRange(0.toULong(), currentText.length.toULong()), newValue
+                )
+            } ?: true
             if (shouldApplyChange) {
                 textField.text = newValue
             }
