+++ widgets/src/iosMain/kotlin/dev/icerock/moko/widgets/core/factory/ButtonWithIconViewFactory.kt
@@ -73,9 +73,9 @@ actual class ButtonWithIconViewFactory actual constructor(
             }
         }
 
-        when (val content = widget.content) {
+        when (widget.content) {
             is ButtonWidget.Content.Text -> {
-                content.text.bind { text ->
+                widget.content.text.bind { text ->
                     val localizedText = text?.localized()
                     val processedText = if (isAllCaps == true) {
                         localizedText?.toUpperCase()
@@ -127,12 +127,12 @@ actual class ButtonWithIconViewFactory actual constructor(
         button.displayLink {
             val icPadding: Double = iconPadding?.toDouble() ?: 0.0
 
-            val newButtonWidth = button.bounds().useContents { this.size.width }
+            val newButtonWidth = button.bounds.useContents { this.size.width }
             if (buttonWidth == newButtonWidth) return@displayLink
             buttonWidth = newButtonWidth
 
-            val iconWidth = button.imageView?.frame()?.useContents { this.size.width } ?: 0.0
-            val titleWidth = button.titleLabel?.frame()?.useContents { this.size.width } ?: 0.0
+            val iconWidth = button.imageView?.frame?.useContents { this.size.width } ?: 0.0
+            val titleWidth = button.titleLabel?.frame?.useContents { this.size.width } ?: 0.0
 
             val paddingTop = padding?.top?.toDouble() ?: 0.0
             var paddingLeft = padding?.start?.toDouble() ?: 0.0
