+++ widgets/src/iosMain/kotlin/dev/icerock/moko/widgets/core/factory/ClickableViewFactory.kt
@@ -21,7 +21,6 @@ actual class ClickableViewFactory actual constructor(
         size: WS,
         viewFactoryContext: ViewFactoryContext
     ): ViewBundle<WS> {
-        @Suppress("UNCHECKED_CAST")
         val childViewBundle =
             widget.child.buildView(viewFactoryContext) as ViewBundle<WS>
