+++ widgets/src/iosMain/kotlin/dev/icerock/moko/widgets/core/screen/sendEmailExt.kt
@@ -4,8 +4,7 @@
 
 package dev.icerock.moko.widgets.core.screen
 
-import dev.icerock.moko.widgets.core.associatedObject
-import dev.icerock.moko.widgets.core.utils.DismissMailControllerDelegate
+import dev.icerock.moko.widgets.core.objc.setAssociatedObject
 import platform.Foundation.NSError
 import platform.MessageUI.MFMailComposeResult
 import platform.MessageUI.MFMailComposeViewController
@@ -23,12 +22,22 @@ actual fun Screen<*>.sendEmail(
     }
 
     val mail = MFMailComposeViewController()
-    val manager = DismissMailControllerDelegate()
+    val manager = EmailManager()
     mail.mailComposeDelegate = manager
-    mail.associatedObject = manager
+    setAssociatedObject(mail, manager)
     mail.setToRecipients(listOf(email))
     mail.setSubject(subject)
     mail.setMessageBody(body, false)
 
     viewController.presentViewController(mail, true, null)
+}
+
+private class EmailManager : NSObject(), MFMailComposeViewControllerDelegateProtocol {
+    override fun mailComposeController(
+        controller: MFMailComposeViewController,
+        didFinishWithResult: MFMailComposeResult,
+        error: NSError?
+    ) {
+        controller.dismissViewControllerAnimated(true, null)
+    }
 }
